requirements.txt
jupyter==1.0.0
numpy==1.19.5
pandas==1.0.3
matplotlib==3.2.1
Pillow==7.1.2
scikit-learn==0.22.2.post1
torch==1.10.2
torchvision==0.11.3
h5py==3.1.0
scikit-image==0.18.1
opencv-python==4.2.0.34
seaborn==0.11.1
scipy==1.4.1
efficientnet==1.1.1
tqdm==4.61.1
lime==0.2.0.1
torchmetrics==0.9.1